swatch bharat new fields addition

+++ rainmaker-pgr/src/main/java/org/egov/pgr/model/SwatchBharat.java
@@ -39,7 +39,7 @@ public class SwatchBharat {
 	private String username;
 	
 	@JsonProperty("workbookid")
-	private int workbookid;
+	private Long workbookid;
 
 	@JsonProperty("auditDetails")
 	private AuditDetails auditDetails;
